latest

+++ src/extensions/logarithm_base2_as_usize.rs
@@ -3,7 +3,7 @@
 
 
 #[inline(always)]
-pub(crate) const fn logarithm_base2_as_usize(value: u32) -> usize
+pub(crate) const fn logarithm_base2_as_usize(value: usize) -> usize
 {
 	value.trailing_zeros() as usize
 }
